Break-up FramePublisher plugin into UI and controller
